+++ Makefile
@@ -3,17 +3,22 @@ CFLAGS = $(foreach dir, $(wildcard inc/*), -I$(dir)) `pkg-config --cflags --libs
 
 #local path from makefile to cfitsio library
 CFITSIO_LIBPATH = lib/cfitsio-4.4.1
-LDFLAGS = $(foreach dir, $(wildcard lib/*), -L$(dir)) `pkg-config --cflags --libs gtk+-3.0` -lcfitsio -lm
-LDFLAGS += -Wl,-rpath,$(CFITSIO_LIBPATH)
+LDFLAGS = -Wl,-rpath,$(CFITSIO_LIBPATH)
+LDFLAGS += $(foreach dir, $(wildcard lib/*), -L$(dir)) `pkg-config --cflags --libs gtk+-3.0` -lcfitsio -lm
 
-SRCDIR = src
-OBJDIR = obj
+SRC_DIR = src
+UI_DIR = $(SRC_DIR)/ui
+HELPERS_DIR = $(SRC_DIR)/helpers
 
-SOURCES = $(wildcard $(SRCDIR)/*.c)
-OBJECTS = $(patsubst $(SRCDIR)/%.c, $(OBJDIR)/%.o, $(SOURCES))
+UI_SRCS = $(wildcard $(UI_DIR)/*.c)
+HELPERS_SRCS = $(wildcard $(HELPERS_DIR)/*.c)
+SOURCES = $(wildcard $(SRC_DIR)/*.c) $(HELPERS_SRCS) $(UI_SRCS)
 
-LIBDIRS = $(wildcard lib/*)
-LIBSA = $(foreach dir, $(LIBDIRS), $(dir)/*.a)
+OBJ_DIR = obj
+OBJECTS = $(patsubst $(SRC_DIR)/%.c, $(OBJ_DIR)/%.o, $(SOURCES))
+
+LIB_DIRS = $(wildcard lib/*)
+LIBS_A = $(foreach dir, $(LIB_DIRS), $(dir)/*.a)
 
 PROGRAM = Asteria
 
@@ -22,7 +27,8 @@ all: $(PROGRAM)
 $(PROGRAM): $(OBJECTS) 
 	$(CC) $^ -o $@ $(LDFLAGS)
 
-$(OBJDIR)/%.o: $(SRCDIR)/%.c
+$(OBJ_DIR)/%.o: $(SRC_DIR)/%.c
+	@mkdir -p $(@D)
 	$(CC) $(CFLAGS) -c $< -o $@
 
 clean:
