+++ src/helpers/app_glib_helper.c
