+++ src/helpers/app_gtk_helper.c
